requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
xgboost==1.7.6
joblib==1.3.2
pyyaml==6.0.1
python-dateutil==2.8.2
requests==2.31.0
pytest==7.4.0
pytest-cov==4.1.0
gunicorn==21.2.0
pandas>=1.5.0
numpy>=1.26.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
tensorflow>=2.20.0
seaborn>=0.12.0
scikit-learn>=1.2.0
scipy>=1.10.0
